requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-appconf==1.0.3
django-cors-headers==3.1.1
django-jwt-auth==0.0.2
django-oauth-toolkit==1.2.0
django-staticfiles==1.2.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-nested-routers==0.91
gunicorn==19.7.1
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
